Move decorators and mixins to spyder.api.config
