+++ spyder/api/config/decorators.py
